requirements.txt
asgiref==3.4.1
branca==0.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.6
django-crispy-forms==1.12.0
django-floppyforms==1.9.0
folium==0.12.1
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
oauthlib==3.1.1
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
